+++ Frontend/src/Components/Home/Home.jsx
@@ -1,16 +1,21 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from './AuthContext';
+import { useAuth } from '../Contexts/AuthContext';
 import './Home.css';
 
 const Home = () => {
   const navigate = useNavigate();
-  const { isAuthenticated, user } = useAuth();
+  const { isAuthenticated, user, logout} = useAuth();
 
   const handleShopNowClick = () => {
     if (!isAuthenticated) {
       navigate('/login');
     }
+  };
+
+  const handleLogout = () => {
+    logout();
+    navigate('/');
   };
 
   return (
@@ -21,6 +26,7 @@ const Home = () => {
           <div className="userProfile">
             <h3>Hi @{user.username}</h3>
             <img src={user.profilePicture || 'default-profile.png'} alt="Profile" className="profilePicture" />
+            <button onClick={handleLogout} className="logoutButton">Logout</button>
           </div>
         )}
       </header>
@@ -35,4 +41,3 @@ const Home = () => {
 };
 
 export default Home;
-
