+++ Frontend/src/Components/Login/LogIn.jsx
@@ -1,52 +1,57 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { Link } from 'react-router-dom';
-import './SignUp.css';
+import { useAuth } from '../Contexts/AuthContext';
+import './Login.css';
 
-const backendApi = import.meta.env.VITE_BACKEND_ADDRESS
+const backendApi = import.meta.env.VITE_BACKEND_ADDRESS;
 
-const Signup = () => {
+const LogIn = () => {
   const [username, setUsername] = useState('');
   const [password, setPassword] = useState('');
   const [error, setError] = useState('');
+  const { login } = useAuth();
   const navigate = useNavigate();
 
-  const handleSignup = async (e) => {
+  const handleLogin = async (e) => {
     e.preventDefault();
     try {
-      const response = await fetch(`${backendApi}/register`, {
+      const response = await fetch(`${backendApi}/login`, {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
         },
+        credentials: 'include',
         body: JSON.stringify({ username, password }),
       });
 
       if (response.ok) {
-        navigate('/'); // Redirect to login page after successful signup
+        const { token, user } = await response.json();
+        login(user);
+        navigate('/');
       } else {
-        setError('User already exists. Please try another username or log in');
+        setError('Invalid credentials. Please try again.');
       }
     } catch (err) {
+      console.log(err);
       setError('Something went wrong. Please try again.');
     }
   };
 
   return (
-    <div className="signupContainer">
-      <header className="signupHeader">
+    <div className="loginContainer">
+      <header className="loginHeader">
         <div className="appName">World Collection</div>
         <div className="authLinks">
           <a href="/login">Log in</a>
           <a href="/register">Sign up</a>
         </div>
       </header>
-      <div className="signupBody">
+      <div className="loginBody">
         <div className="circle"></div>
-        <div className="loginPrompt">
-          Already have an account? <Link to={"/login"} className="viewButton" >Log in</Link>
+        <div className="signupPrompt">
+          Don’t have an account? <a href="/register" className="signupLink">Sign up</a>
         </div>
-        <form className="signupForm" onSubmit={handleSignup}>
+        <form className="loginForm" onSubmit={handleLogin}>
           <input
             type="text"
             placeholder="Username"
@@ -62,11 +67,11 @@ const Signup = () => {
             onChange={(e) => setPassword(e.target.value)}
           />
           {error && <p className="error">{error}</p>}
-          <button type="submit" className="signupButton">Sign up</button>
+          <button type="submit" className="loginButton">Log in</button>
         </form>
       </div>
     </div>
   );
 };
 
-export default Signup;
+export default LogIn;
